+++ crates/base/core/src/vita/play/mod.rs
@@ -9,18 +9,18 @@
 
 // pub mod comp; // computer and simulated games
 // pub mod dance; // expressive movement
-// pub mod game; // games across cultures
+pub mod game; // games across cultures
 // pub mod music;// music, rhythm, embodied sound
 // pub mod sport; // play-oriented physical competition
 
 crate::structural_mods! { // _pub_mods
     _pub_mods {
-        // pub use super::{
-        //     comp::_all::*,
-        //     dance::_all::*,
-        //     game::_all::*,
-        //     music::_all::*,
-        //     sport::_all::*,
-        // };
+        pub use super::{
+            // comp::_all::*,
+            // dance::_all::*,
+            game::_all::*,
+            // music::_all::*,
+            // sport::_all::*,
+        };
     }
 }