+++ src/vita/mod.rs
@@ -1,8 +1,5 @@
 // devela::vita
 //
-#![cfg_attr(not(feature = "__docs_internal"), doc(hidden))]
-//
-#![doc = crate::_tags!(wip)]
 #![doc = crate::_DOC_VITA!()] // public, root
 #![doc = crate::_DOC_VITA_MODULES!()]
 #![doc = crate::_doc!(flat:"vita")]
@@ -12,7 +9,7 @@
 #![cfg_attr(feature = "safe_vita", forbid(unsafe_code))]
 // docs
 crate::CONST! { pub(crate) _DOC_VITA_MODULES =
-    crate::_doc!(modules: crate; vita: _); // body, craft, home, love, mind, play
+    crate::_doc!(modules: crate; vita: play); // body, craft, home, love, mind
 }
 
 // pub mod body; // Embodied capability, health, and physical limits of living beings.
@@ -20,18 +17,18 @@ crate::CONST! { pub(crate) _DOC_VITA_MODULES =
 // pub mod home; // Inhabited space, from dwelling to built and shared environments.
 // pub mod love; // Relational life, from kinship and care to intimacy and community.
 // pub mod mind; // Cognition, memory, meaning, and inner orientation of lived experience.
-// pub mod play; // Expression, play, and shared enjoyment beyond necessity or survival.
+pub mod play; // Expression, play, and shared enjoyment beyond necessity or survival.
 
 crate::structural_mods! { // _pub_mods, _crate_internals
     _pub_mods {
-        // pub use super::{
-        //     body::_all::*,
-        //     craft::_all::*,
-        //     home::_all::*,
-        //     love::_all::*,
-        //     mind::_all::*,
-        //     play::_all::*,
-        // };
+        pub use super::{
+            // body::_all::*,
+            // craft::_all::*,
+            // home::_all::*,
+            // love::_all::*,
+            // mind::_all::*,
+            play::_all::*,
+        };
     }
     _crate_internals {
         pub(crate) use super::_DOC_VITA_MODULES;